+++ packages/common/src/relay/calling/Call.ts
@@ -1,8 +1,8 @@
 import { v4 as uuidv4 } from 'uuid'
 import logger from '../../util/logger'
 import { Execute } from '../../messages/Blade'
-import { CallState, DisconnectReason, DEFAULT_CALL_TIMEOUT, CallNotification, CallRecordState, CallPlayState, CallPromptState, CallConnectState, CALL_STATES, CallFaxState } from '../../util/constants/relay'
-import { ICall, ICallOptions, ICallDevice, IMakeCallParams, ICallingPlay, ICallingCollect, DeepArray } from '../../util/interfaces'
+import { CallState, DisconnectReason, DEFAULT_CALL_TIMEOUT, CallNotification, CallRecordState, CallPlayState, CallPromptState, CallConnectState, CALL_STATES, CallFaxState, CallDetectState, CallDetectType } from '../../util/constants/relay'
+import { ICall, ICallOptions, ICallDevice, IMakeCallParams, ICallingPlay, ICallingCollect, DeepArray, ICallingDetect } from '../../util/interfaces'
 import { reduceConnectParams } from '../helpers'
 import Calling from './Calling'
 import { isFunction } from '../../util/helpers'
@@ -30,6 +30,9 @@ import FaxReceive from './components/FaxReceive'
 import FaxResult from './results/FaxResult'
 import FaxAction from './actions/FaxAction'
 import FaxSend from './components/FaxSend'
+import Detect from './components/Detect'
+import DetectResult from './results/DetectResult'
+import DetectAction from './actions/DetectAction'
 
 export default class Call implements ICall {
   public id: string
@@ -308,6 +311,111 @@ export default class Call implements ICall {
     return new FaxAction(component)
   }
 
+  async detect(type: string, params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectResult> {
+    const detect: ICallingDetect = { type, params }
+    const component = new Detect(this, detect, timeout)
+    this._addComponent(component)
+    await component._waitFor(CallDetectState.Error, CallDetectState.Finished)
+
+    return new DetectResult(component)
+  }
+
+  async detectAsync(type: string, params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectAction> {
+    const detect: ICallingDetect = { type, params }
+    const component = new Detect(this, detect, timeout)
+    this._addComponent(component)
+    await component.execute()
+
+    return new DetectAction(component)
+  }
+
+  async detectHuman(params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectResult> {
+    const detect: ICallingDetect = { type: CallDetectType.Machine, params }
+    const component = new Detect(this, detect, timeout)
+    this._addComponent(component)
+    await component._waitFor(CallDetectState.Human, CallDetectState.Error, CallDetectState.Finished)
+
+    return new DetectResult(component)
+  }
+
+  async detectHumanAsync(params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectAction> {
+    const detect: ICallingDetect = { type: CallDetectType.Machine, params }
+    const component = new Detect(this, detect, timeout)
+    component.eventsToWait = [CallDetectState.Human, CallDetectState.Error, CallDetectState.Finished]
+    this._addComponent(component)
+    await component.execute()
+
+    return new DetectAction(component)
+  }
+
+  async detectMachine(params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectResult> {
+    const detect: ICallingDetect = { type: CallDetectType.Machine, params }
+    const component = new Detect(this, detect, timeout)
+    this._addComponent(component)
+    await component._waitFor(CallDetectState.Machine, CallDetectState.Ready, CallDetectState.NotReady, CallDetectState.Error, CallDetectState.Finished)
+
+    return new DetectResult(component)
+  }
+
+  async detectMachineAsync(params: ICallingDetect['params'] = {}, timeout: number = null): Promise<DetectAction> {
+    const detect: ICallingDetect = { type: CallDetectType.Machine, params }
+    const component = new Detect(this, detect, timeout)
+    component.eventsToWait = [CallDetectState.Machine, CallDetectState.Ready, CallDetectState.NotReady, CallDetectState.Error, CallDetectState.Finished]
+    this._addComponent(component)
+    await component.execute()
+
+    return new DetectAction(component)
+  }
+
+  async detectFax(tone: string = null, timeout: number = null): Promise<DetectResult> {
+    const faxEvents: string[] = [CallDetectState.CED, CallDetectState.CNG]
+    let events: string[] = [CallDetectState.Error, CallDetectState.Finished]
+    const params: { tone?: string } = {}
+    if (tone && faxEvents.includes(tone)) {
+      params.tone = tone
+      events.push(tone)
+    } else {
+      events = events.concat(faxEvents)
+    }
+
+    const detect: ICallingDetect = { type: CallDetectType.Fax, params }
+    const component = new Detect(this, detect, timeout)
+    this._addComponent(component)
+    await component._waitFor(...events)
+
+    return new DetectResult(component)
+  }
+
+  async detectFaxAsync(tone: string = null, timeout: number = null): Promise<DetectAction> {
+    const faxEvents: string[] = [CallDetectState.CED, CallDetectState.CNG]
+    let events: string[] = [CallDetectState.Error, CallDetectState.Finished]
+    const params: { tone?: string } = {}
+    if (tone && faxEvents.includes(tone)) {
+      params.tone = tone
+      events.push(tone)
+    } else {
+      events = events.concat(faxEvents)
+    }
+
+    const detect: ICallingDetect = { type: CallDetectType.Fax, params }
+    const component = new Detect(this, detect, timeout)
+    component.eventsToWait = events
+    this._addComponent(component)
+    await component.execute()
+
+    return new DetectAction(component)
+  }
+
+  async detectDigit(digits: string = null, timeout: number = null): Promise<DetectResult> {
+    const params = digits ? { digits } : {}
+    return this.detect(CallDetectType.Digit, params, timeout)
+  }
+
+  async detectDigitAsync(digits: string = null, timeout: number = null): Promise<DetectAction> {
+    const params = digits ? { digits } : {}
+    return this.detectAsync(CallDetectType.Digit, params, timeout)
+  }
+
   /**
    * Registers a callback to dispatch when the 'event' occur.
    * @param event - Event to listen to.
@@ -374,6 +482,17 @@ export default class Call implements ICall {
     this._dispatchCallback('fax.stateChange', params)
     if (params.fax && params.fax.type) {
       this._dispatchCallback(`fax.${params.fax.type}`, params)
+    }
+  }
+
+  _detectChange(params: any) {
+    this._notifyComponents(CallNotification.Detect, params.control_id, params)
+
+    const { params: { event = null } } = params.detect
+    if (event === CallDetectState.Finished || event === CallDetectState.Error) {
+      this._dispatchCallback(`detect.${event}`, params)
+    } else if (event) {
+      this._dispatchCallback('detect.update', params)
     }
   }